# backend/openapi/Castor.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Castor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/innovations": {
      "post": {
        "tags": [
          "AdminInnovations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInnovationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInnovationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInnovationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/innovations/{innovationId}": {
      "put": {
        "tags": [
          "AdminInnovations"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseInnovationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseInnovationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseInnovationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/innovations/{innovationId}": {
      "get": {
        "tags": [
          "Innovations"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/genomes": {
      "get": {
        "tags": [
          "InnovationGenomes"
        ],
        "parameters": [
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InnovationGenomeResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InnovationGenomeResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InnovationGenomeResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/genomes/{genomeId}": {
      "get": {
        "tags": [
          "InnovationGenomes"
        ],
        "parameters": [
          {
            "name": "genomeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "InnovationGenomes"
        ],
        "parameters": [
          {
            "name": "genomeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseInnovationGenomeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseInnovationGenomeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseInnovationGenomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/genomes/{genomeId}/approve": {
      "post": {
        "tags": [
          "InnovationGenomes"
        ],
        "parameters": [
          {
            "name": "genomeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InnovationGenomeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/genomes/{genomeId}/recalculate": {
      "post": {
        "tags": [
          "InnovationGenomes"
        ],
        "parameters": [
          {
            "name": "genomeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInnovationRequest": {
        "required": [
          "title",
          "shortDescription",
          "categories",
          "solution",
          "problems",
          "targetGroup",
          "beneficiaries",
          "evidence",
          "organization",
          "cardUrl",
          "videoUrl",
          "materialsZipUrl",
          "cardPdfUrl",
          "termsUrl",
          "stage"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "shortDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "categories": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "solution": {
            "type": [
              "null",
              "string"
            ]
          },
          "problems": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetGroup": {
            "type": [
              "null",
              "string"
            ]
          },
          "beneficiaries": {
            "type": [
              "null",
              "string"
            ]
          },
          "evidence": {
            "type": [
              "null",
              "string"
            ]
          },
          "organization": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "videoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "materialsZipUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardPdfUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "termsUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "stage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "InnovationGenomeResponse": {
        "required": [
          "id",
          "innovationId",
          "innovationTitle",
          "status",
          "rootCauses",
          "mechanisms",
          "targetGroups",
          "requiredInstitutions",
          "requiredPeople",
          "requiredBudget",
          "requiredInfrastructure",
          "scale",
          "challengeAreaCodes",
          "summary",
          "approvedAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "innovationId": {
            "type": "string",
            "format": "uuid"
          },
          "innovationTitle": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "rootCauses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mechanisms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetGroups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requiredInstitutions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requiredPeople": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requiredBudget": {
            "type": [
              "null",
              "string"
            ]
          },
          "requiredInfrastructure": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "scale": {
            "type": [
              "null",
              "string"
            ]
          },
          "challengeAreaCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "summary": {
            "type": "string"
          },
          "approvedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InnovationResponse": {
        "required": [
          "id",
          "title",
          "shortDescription",
          "categories",
          "solution",
          "problems",
          "targetGroup",
          "beneficiaries",
          "evidence",
          "organization",
          "cardUrl",
          "videoUrl",
          "materialsZipUrl",
          "cardPdfUrl",
          "termsUrl",
          "featured",
          "inServiceModel",
          "stage",
          "challengeAreaCodes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "shortDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "solution": {
            "type": [
              "null",
              "string"
            ]
          },
          "problems": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetGroup": {
            "type": [
              "null",
              "string"
            ]
          },
          "beneficiaries": {
            "type": [
              "null",
              "string"
            ]
          },
          "evidence": {
            "type": [
              "null",
              "string"
            ]
          },
          "organization": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "videoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "materialsZipUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardPdfUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "termsUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "featured": {
            "type": "boolean"
          },
          "inServiceModel": {
            "type": "boolean"
          },
          "stage": {
            "type": "string"
          },
          "challengeAreaCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ReviseInnovationGenomeRequest": {
        "required": [
          "rootCauses",
          "mechanisms",
          "targetGroups",
          "requiredInstitutions",
          "requiredPeople",
          "requiredBudget",
          "requiredInfrastructure",
          "scale",
          "challengeAreaCodes",
          "summary"
        ],
        "type": "object",
        "properties": {
          "rootCauses": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "mechanisms": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "targetGroups": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "requiredInstitutions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "requiredPeople": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "requiredBudget": {
            "type": [
              "null",
              "string"
            ]
          },
          "requiredInfrastructure": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "scale": {
            "type": [
              "null",
              "string"
            ]
          },
          "challengeAreaCodes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ReviseInnovationRequest": {
        "required": [
          "title",
          "shortDescription",
          "categories",
          "solution",
          "problems",
          "targetGroup",
          "beneficiaries",
          "evidence",
          "organization",
          "cardUrl",
          "videoUrl",
          "materialsZipUrl",
          "cardPdfUrl",
          "termsUrl",
          "stage"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "shortDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "categories": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "solution": {
            "type": [
              "null",
              "string"
            ]
          },
          "problems": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetGroup": {
            "type": [
              "null",
              "string"
            ]
          },
          "beneficiaries": {
            "type": [
              "null",
              "string"
            ]
          },
          "evidence": {
            "type": [
              "null",
              "string"
            ]
          },
          "organization": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "videoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "materialsZipUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardPdfUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "termsUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "stage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}